refactor: update imports to use mcp-manager for server service
