+++ apps/electron/src/main/modules/mcp-core/mcp-manager/server-service.ts
