+++ apps/electron/src/main/modules/workspace/platform-api-manager.ts
@@ -17,7 +17,7 @@ import {
   resetTokenRepository,
   resetWorkspaceRepository,
 } from "../../infrastructure/database";
-import { ServerService } from "@/main/modules/mcp-core/server/server-service";
+import { ServerService } from "@/main/modules/mcp-core/mcp-manager/server-service";
 import { McpAppsService } from "@/main/modules/mcp-core/apps/mcp-apps-service";
 import { LogService } from "../mcp-core/log/log-service";
 import { SettingsService } from "../settings/settings.service";
